Backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.8.1
supabase==2.13.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.28.1
email-validator==2.2.0
